# dev/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Domovoy API
  version: 0.1.0
  description: Персональный игровой слой поверх X5 Клуба. Auth отсутствует, пользователь демо задаётся user_id.
servers:
  - url: http://localhost:8000
...
paths:
...
  /api/v1/users:
    get:
      tags: [users]
      operationId: listUsers
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 500 }
      responses:
        "200":
          description: Список демо-пользователей
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /api/v1/users/{user_id}/home:
    get:
      tags: [users]
      operationId: getHome
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Агрегат для экрана Home
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    UserId:
      in: path
      name: user_id
      required: true
      schema: { type: integer }
  responses:
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
    UserSummary:
      type: object
      required: [id, pseudonym, segment, level]
      properties:
        id: { type: integer }
        pseudonym: { type: string }
        segment: { type: string, enum: [regular_mid, light, heavy, dormant] }
        level: { type: integer }
    UserListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/UserSummary" }
    DomovoyState:
      type: object
      required: [xp, level, xp_to_next_level, mood, mood_reason, streak_weeks, items]
      properties:
        xp: { type: integer }
        level: { type: integer }
        xp_to_next_level: { type: integer }
        mood: { type: string, enum: [cheerful, cozy, healthy, bored, sleepy] }
        mood_reason: { type: string }
        streak_weeks: { type: integer }
        items:
          type: array
          items: { type: string }
    SavingsCategory:
      type: object
      required: [category, amount, items_count, top_products]
      properties:
        category: { type: string }
        amount: { type: number }
        items_count: { type: integer }
        top_products:
          type: array
          items: { type: string }
    SavingsSummary:
      type: object
      required: [period, amount, previous_amount, delta, discount_amount, points_earned, points_spent, receipts_count, top_categories]
      properties:
        period: { type: string, enum: [week, month] }
        amount: { type: number }
        previous_amount: { type: number }
        delta: { type: number }
        discount_amount: { type: number }
        points_earned: { type: integer }
        points_spent: { type: integer }
        receipts_count: { type: integer }
        top_categories:
          type: array
          items: { $ref: "#/components/schemas/SavingsCategory" }
...
    ChallengeEconomics:
      type: object
      required: [avg_basket, expected_incremental_purchases, expected_incremental_margin, max_reward_rub, contribution_margin, reward_share_max]
      properties:
        avg_basket: { type: number }
        expected_incremental_purchases: { type: number }
        expected_incremental_margin: { type: number }
        max_reward_rub: { type: number }
        contribution_margin: { type: number }
        reward_share_max: { type: number }
    Challenge:
      type: object
      required: [id, type, category, status, is_hero, baseline, target, progress, period_start, period_end, reward_xp, reward_points, title, body]
      properties:
        id: { type: integer }
        type: { type: string, enum: [frequency, category] }
        category: { type: [string, "null"] }
        status: { type: string, enum: [active, completed, failed, expired] }
        is_hero: { type: boolean }
        baseline: { type: number }
        target: { type: number }
        progress: { type: number }
        period_start: { type: string, format: date-time }
        period_end: { type: string, format: date-time }
        reward_xp: { type: integer }
        reward_points: { type: integer }
        title: { type: string }
        body: { type: string }
    ChallengeDetail:
      allOf:
        - $ref: "#/components/schemas/Challenge"
        - type: object
          required: [explanation, rationale_features, economics, copy_source]
          properties:
            explanation: { type: string }
            rationale_features:
              type: object
              additionalProperties: true
            economics: { $ref: "#/components/schemas/ChallengeEconomics" }
            copy_source: { type: string, enum: [llm, template] }
...
    LeagueTeaser:
      type: object
      required: [division, rank, size, zone]
      properties:
        division: { type: integer }
        rank: { type: integer }
        size: { type: integer }
        zone: { type: string, enum: [promotion, safe, demotion] }
    ReferralTeaser:
      type: object
      required: [code, invited_count, rewarded_count]
      properties:
        code: { type: string }
        invited_count: { type: integer }
        rewarded_count: { type: integer }
    HomeResponse:
      type: object
      required: [user, domovoy, savings, points_balance, insight, hero_challenge, league, referral, recommended_mechanic]
      properties:
        user: { $ref: "#/components/schemas/UserSummary" }
        domovoy: { $ref: "#/components/schemas/DomovoyState" }
        savings: { $ref: "#/components/schemas/SavingsSummary" }
        points_balance: { type: integer }
        insight: { type: string }
        hero_challenge:
          oneOf:
            - $ref: "#/components/schemas/ChallengeDetail"
            - type: "null"
        league:
          oneOf:
            - $ref: "#/components/schemas/LeagueTeaser"
            - type: "null"
        referral: { $ref: "#/components/schemas/ReferralTeaser" }
        recommended_mechanic:
          type: object
          required: [mechanic, reason]
          properties:
            mechanic: { type: string, enum: [challenge, league, referral] }
            reason: { type: string }
...